@Baunsgaard 2023-01-06
@mboehm7 ...
Does it make sense to make memory estimation parallel? or should it be simplified to not be exact in frame cases?