@Baunsgaard 2023-01-06
Time if ---wrongly--- FrameBlocks always return they are 4KB on calls to estimate , here apply schema goes from 1.2 to 0.01 sec on calls to getInMemorySize.

My profiler similarly says that getUTFSize in writing frame blocks to disk take significantly much time.

```txt
ystemDS Statistics:
Total elapsed time:		4.530 sec.
Total compilation time:		0.237 sec.
Total execution time:		4.292 sec.
Cache hits (Mem/Li/WB/FS/HDFS):	4/0/0/0/1.
Cache writes (Li/WB/FS/HDFS):	0/0/0/0.
Cache times (ACQr/m, RLS, EXP):	1.613/0.000/0.000/0.876 sec.
HOP DAGs recompiled (PRED, SB):	0/1.
HOP DAGs recompile time:	0.003 sec.
Total JIT compile time:		4.22 sec.
Total JVM GC count:		6.
Total JVM GC time:		0.878 sec.
Heavy hitter instructions:
 #  Instruction   Time(s)  Count
 1  detectSchema    3.386      1
 2  write           0.876      1
 3  applySchema     0.014      1
 4  createvar       0.008      4
```