@mboehm7 2023-01-06
no, this is the wrong thing to optimize - instead you might push it into the producer, the read here and simply materialize the size for string (variable size) columns.